2.10 release
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@269 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ SettingsManager.h
@@ -329,7 +329,7 @@ public:
 			return false;
 		Lock l(cs);
 
-		if (searchHistory.size() > (getInstance()->get(SEARCH_HISTORY)))
+		if ((int)searchHistory.size() > (getInstance()->get(SEARCH_HISTORY)))
 				searchHistory.erase(searchHistory.begin());
 
 		searchHistory.push_back(search);
